Fix coverage workflow (#1939)

+++ tools/test-op.sh
@@ -4,25 +4,23 @@ set -e
 
 PR_ID=$1
 
-# Replace "__ALL__" with all tests
+# Leave this for debugging's purpose
+echo "PR_ID=${PR_ID}"
+
+COLLECT_COVERAGE=""
+
 if [[ "$CHANGED_FILES" == "__ALL__" ]]; then
-  # Temporay hack
-  CHANGED_FILES=(
-    "tests/test_tensor_constructor_ops.py"
-  )
-  # CHANGED_FILES=$(find tests -name "test*.py")
-  # for full-range tests, generate summary report
+  # Replace "__ALL__" with all tests
+  CHANGED_FILES=$(find tests -name "test*.py")
+  # add options to generate summary report
   EXTRA_OPTS="--md-report --md-report-output=${PR_ID}-summary.md"
-  echo "TIMESTAMP=${PR_ID}"
   SUFFIX=""
+  COLLECT_COVERAGE="yes"
 else
   # for per-PR test, fail early
   EXTRA_OPTS="-x"
-  echo "PR_ID=${PR_ID}"
   SUFFIX="-${GITHUB_SHA::7}"
 fi
-
-
 
 # Test cases that needs to run quick cpu tests
 QUICK_CPU_TESTS=(
@@ -48,11 +46,12 @@ for item in $CHANGED_FILES; do
       # skip DSA test for now
       ;;
     tests/test_quant.py)
-      # skip
+      # skip because it always fail
       ;;
     tests/*) TEST_CASES+=($item)
   esac
 
+  # filter out tests that do not need quick CPU mode tests
   for item_cpu in "${QUICK_CPU_TESTS[@]}"; do
     if [[ "$item" == "$item_cpu" ]]; then
       TEST_CASES_CPU+=($item)
@@ -82,7 +81,7 @@ fi
 
 # Process coverage data only when full-range testing
 # Coverage data HTML dumped to `htmlcov/` by default
-# if [[ "$CHANGED_FILES" == "__ALL__" ]]; then
+if [ -n "$COLLECT_COVERAGE" ]; then
   coverage combine
   coverage html
   rm -fr coverage
@@ -90,4 +89,4 @@ fi
   mv htmlcov coverage/
   echo "${PR_ID}${SUFFIX::7}" > coverage/COVERAGE_ID
   mv ${PR_ID}-summary.md coverage/
-# fi
+fi
